fixed updating PB for real Centralized stats in statistics.go added helper function to prevent invalid commands for keybinds
